+++ 01_krepios/init.lua
@@ -82,6 +82,7 @@ local myPackages = {
     "01_krepios/upgrades.lua",
     "01_krepios/flying_trader.lua",
     "01_krepios/chatter.lua",
+    "01_krepios/raid.lua",
 
     "01_krepios/upgrades/jump_calculator.lua",
     "01_krepios/upgrades/laser_refit.lua",