+++ 01_krepios/lang/de/init.lua
@@ -15,6 +15,7 @@ local myPackages = {
     "01_krepios/lang/de/de_graveyard.lua",
     "01_krepios/lang/de/de_nebulas.lua",
     "lang/de/de_player.lua",
+    "01_krepios/lang/de/de_pirates.lua",
     "01_krepios/lang/de/de_player.lua",
     "01_krepios/lang/de/de_player_science.lua",
     "lang/de/de_products.lua",