+++ 01_krepios/lang/en/en_drops.lua
@@ -12,23 +12,21 @@ My.Translator:register("en", {
         })
     end,
 
-    drops_description_unknown = "unknown object",
+    drops_generic_description_full = "Contents:",
 
-    drops_description_full = function(shipCallSign, value, energy)
+    drops_description_full = function(shipCallSign)
         return Util.random({
             f("This shipping container originates from the ship %s that crashed in this minefield.", shipCallSign),
             f("The container seems to have been lost by the ship %s.", shipCallSign),
             f("The container comes from the damaged ship %s.", shipCallSign),
         }) .. " " .. Util.random({
-            f("It contains %0.2fRP and %d energy.", value, energy),
-            f("In it there are %0.2fRP and %d energy.", value, energy),
-            f("The scans show that its contents are %0.2fRP and %d energy.", value, energy),
+            "It contains:",
+            "In it there is: ",
+            "The scans show that its contents are:",
         })
     end,
 
-    drops_pickup = function(value, energy)
-        return f("You have collected %0.2fRP and %d energy.", value, energy)
-    end,
+    drops_pickup = "You have collected:",
 
     drops_ship_description = function(shipCallSign)
         return Util.random({
@@ -41,5 +39,9 @@ My.Translator:register("en", {
             "How it ended up in a mine field remains a mystery.",
         })
     end,
+
+    drops_content_energy = "Energy",
+    drops_content_reputation = "RP",
+
 
 })